+++ src/ui/pages/index.html/settings.svelte
@@ -1,46 +1,13 @@
 <script lang="ts">
-  import { readFileAsText, pickFile } from "@/ui/lib/dom";
-  import { rpc } from "@/ui/lib/rpc";
-  import { z } from "zod";
-  import { updateStats } from "@/ui/store/statsStore";
+  import { handleImport } from "@/ui/lib/commands";
 
   let errorMessage = "";
 
-  const dbImportSchema = z.object({
-    document: z.array(z.any()),
-  });
-
   const handleFileUpload = async () => {
     errorMessage = "";
-    try {
-      const file = await pickFile(".json");
-      if (file.type !== "application/json") {
-        errorMessage = "Please upload a JSON file.";
-        return;
-      }
-
-      const text = await readFileAsText(file);
-      const content = JSON.parse(text);
-      const result = dbImportSchema.safeParse(content);
-      if (!result.success) {
-        errorMessage = "Invalid JSON file. Please upload a valid JSON file.";
-        console.error("Error parsing JSON:", result);
-        return;
-      }
-
-      const documents = result.data.document;
-      await rpc(["importDocumentsJSONv1", { document: documents }]);
-      console.log("Imported:", documents.length, "documents");
-
-      // Update stats after successful import
-      await updateStats();
-    } catch (error) {
-      if (error instanceof Error && error.message === "No file selected") {
-        // User cancelled file selection, do nothing
-        return;
-      }
-      errorMessage = "Invalid JSON file. Please upload a valid JSON file.";
-      console.error("Error importing JSON:", error);
+    const result = await handleImport();
+    if (!result.success && result.message) {
+      errorMessage = result.message;
     }
   };
 </script>
